Restrict the user from adding users or doing a factory reset
Adding users was a workaround, since only the current user has their camera
disabled. Doing a factory reset was probably also a workaround. Now, when we
check if the user is a device admin and see that they are, we add those
restrictions.

It looks like DISALLOW_ADD_USER is enough to also disallow the guest user.